+++ memo/src/app/page.tsx
@@ -1,26 +1,24 @@
 "use client";
 
 import { MemoList } from "@/components/MemoList";
-import { ChangeEvent, useState, FC } from "react";
+import { useMemoList } from "@/hooks/useMemoList";
+import { ChangeEvent, useState, FC, useCallback } from "react";
 
 const Home: FC = () => {
+  const { memos, addMemo, deleteMemo} = useMemoList();
+
   const [text, setText] = useState<string>("");
-  const [memos, setMemos] = useState<string[]>([]);
 
   const onChangeText = (e: ChangeEvent<HTMLInputElement>) => setText(e.target.value);
 
   const onClickAdd = () => {
-    const newMemos = [...memos];
-    newMemos.push(text);
-    setMemos(newMemos);
+    addMemo(text);
     setText("");
   };
 
-  const onClickDelete = (index: number) => {
-    const newMemos = [...memos];
-    newMemos.splice(index, 1);
-    setMemos(newMemos);
-  };
+  const onClickDelete = useCallback((index: number) => {
+    deleteMemo(index);
+  }, [deleteMemo]);
 
   return (
     <>
